+++ pages/lp.tsx
@@ -1,0 +1,16 @@
+'use client'
+import { HeroSection } from "@/components/LP/hero-section";
+import { FeaturesSection } from "@/components/LP/feature-section";
+import { HowItWorks } from "@/components/LP/how-it-works";
+import { CTASection } from "@/components/LP/cta-section";
+
+export default function Home() {
+  return (
+    <main className="min-h-screen">
+      <HeroSection />
+      <FeaturesSection />
+      <HowItWorks />
+      <CTASection />
+    </main>
+  );
+}
